CRSF: address code review feedback
- Use uintptr_t for pointer cast in accgyro_mpu.c (64-bit safety)
- Fix baudrate scaling with 64-bit math and divide-by-zero guard
- Add USE_TELEMETRY_CRSF guard to crsfScheduleTelemetryResponse()
- Use G_ACCELERATION from flight/pid.h instead of local define
- Add CRSF_FRAME_BARO_SENSOR_INDEX for non-V3 builds (fix collision)
- Use cmpTimeUs() for timer wrap-around safe comparison
- Clamp DOP values to UINT8_MAX to prevent overflow
- Fix velD negation overflow when value equals INT16_MIN
- Rename timed schedule enum to CRSF_TIMED_FRAME_* prefix

+++ src/main/drivers/accgyro/accgyro_mpu.c
@@ -283,7 +283,7 @@ bool mpuGyroReadSPI(gyroDev_t *gyro)
         if (gyro->detectedEXTI > GYRO_EXTI_DETECT_THRESHOLD) {
 #ifdef USE_DMA
             if (spiUseDMA(&gyro->dev)) {
-                gyro->dev.callbackArg = (uint32_t)gyro;
+                gyro->dev.callbackArg = (uintptr_t)gyro;
                 gyro->dev.txBuf[0] = gyro->dmaReadRegStart | 0x80;
                 gyro->segments[0].len = gyro->gyroDataReg - gyro->dmaReadRegStart + sizeof(uint8_t) + 3 * sizeof(int16_t);  // assumes gyros come last
                 gyro->segments[0].callback = mpuIntCallback;
